Remove unused dependencies, cleanup readme and add new platforms to the list of options
