+++ qa/Gemfile
@@ -3,5 +3,3 @@ gem "runner-tool", :git => "https://github.com/purbon/runner-tool.git"
 gem "rspec", "~> 3.1.0"
 gem "rake"
 gem "pry", :group => :test
-gem "stud"
-gem "logstash-devutils"
