+++ qa/rspec/commands.rb
@@ -5,6 +5,8 @@ require_relative "./commands/redhat"
 require_relative "./commands/suse"
 require_relative "./commands/centos/centos-6"
 require_relative "./commands/oel/oel-6"
+require_relative "./commands/ubuntu/ubuntu-1604"
+require_relative "./commands/suse/sles-11"
 
 require "forwardable"
 
@@ -95,12 +97,20 @@ module ServiceTester
       case type
       when "debian"
         if host.start_with?("ubuntu")
-          return UbuntuCommands.new
+          if host == "ubuntu-1604"
+            return Ubuntu1604Commands.new
+          else
+            return UbuntuCommands.new
+          end
         else
           return DebianCommands.new
         end
       when "suse"
-        return SuseCommands.new
+        if host == "sles-11"
+          return Sles11Commands.new
+        else
+          return SuseCommands.new
+        end
       when "redhat"
         if host == "centos-6"
           return Centos6Commands.new
