+++ qa/rspec/commands/suse.rb
@@ -13,8 +13,8 @@ module ServiceTester
       stdout.match(/^i | logstash | An extensible logging pipeline | package$/)
     end
 
-    def package_for(version)
-      File.join(ServiceTester::Base::LOCATION, "logstash-#{version}.noarch.rpm")
+    def package_for(version, base=ServiceTester::Base::LOCATION)
+      File.join(base, "logstash-#{version}.rpm")
     end
 
     def install(package, host=nil)
